Fixed return value of long_from in the big endian version

+++ include/pico_constants.h
@@ -30,7 +30,7 @@ static inline uint16_t short_from(void *_p)
   return r;
 }
 
-static inline uint16_t long_from(void *_p)
+static inline uint32_t long_from(void *_p)
 {
   unsigned char *p = (unsigned char *)_p;
   uint32_t r = (p[0] << 24) + (p[1] << 16) + (p[2] << 8) + p[3];
